+++ main.go
@@ -1,15 +1,9 @@
 package main
 
 import (
-	"fmt"
-	"os"
-
 	"github.com/elyscape/az-dns/cmd"
 )
 
 func main() {
-	if err := cmd.RootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
+	cmd.Execute()
 }